split client into client and iframe client
