+++ client/client.js
@@ -13,70 +13,52 @@ import Plugins from './components/Plugins';
 import App from './components/App';
 import FourOhFour from './components/FourOhFour';
 import map from 'lodash/map';
-import find from 'lodash/find';
 import createNavigationStore from './navigationStore';
 
-if (window.frameElement) {
-  // in iframe
-  window.$INITIALIZE_COMPONENT_GUI = function initializeComponentGui(components) {
-    const componentData = components[window.COMPONENT_PATH];
-    let Component = componentData.getCompiledComponent();
-    if (Component.default) {
-      Component = Component.default;
-    }
-    const activePlugin = find(componentData.getMeta(), (plugin) => plugin.name === 'react');
+let rootElement;
+window.$INITIALIZE_COMPONENT_GUI = function initializeComponentGui(components) {
+  if (rootElement) {
+    rootElement.forceUpdate();
+    return;
+  }
 
-    ReactDOM.render(
-      (activePlugin.frontendPlugin(Component)),
-      document.getElementById('root')
-    );
-  };
-} else {
-  let rootElement;
-  window.$INITIALIZE_COMPONENT_GUI = function initializeComponentGui(components) {
-    if (rootElement) {
-      rootElement.forceUpdate();
-      return;
-    }
+  const navigationStore = createNavigationStore(components);
 
-    const navigationStore = createNavigationStore(components);
+  // Generate a view per user component that renders the frontend part of the
+  // plugins for each component
+  const routes = map(components, (component, componentPath) => (
+    <Route
+      key={componentPath}
+      path={componentPath}
+      component={() => (
+        <Plugins
+          componentPath={componentPath}
+          componentData={component}
+          navigationStore={navigationStore}
+        />
+      )}
+    />
+  ));
 
-    // Generate a view per user component that renders the frontend part of the
-    // plugins for each component
-    const routes = map(components, (component, componentPath) => (
+  const appRoute = (<Route
+    path="/"
+    component={(props) => (
+      <App {...props} navigationStore={navigationStore} />
+    )}
+  >
+    {routes}
+  </Route>);
+
+  rootElement = ReactDOM.render(
+    (<Router history={hashHistory}>
+      {appRoute}
       <Route
-        key={componentPath}
-        path={componentPath}
-        component={() => (
-          <Plugins
-            componentPath={componentPath}
-            componentData={component}
-            navigationStore={navigationStore}
-          />
+        path="*"
+        component={(props) => (
+          <FourOhFour {...props} navigationStore={navigationStore} />
         )}
       />
-    ));
-
-    const appRoute = (<Route
-      path="/"
-      component={(props) => (
-        <App {...props} navigationStore={navigationStore} />
-      )}
-    >
-      {routes}
-    </Route>);
-
-    rootElement = ReactDOM.render(
-      (<Router history={hashHistory}>
-        {appRoute}
-        <Route
-          path="*"
-          component={(props) => (
-            <FourOhFour {...props} navigationStore={navigationStore} />
-          )}
-        />
-      </Router>),
-      document.getElementById('carte-blanche-root')
-    );
-  };
-}
+    </Router>),
+    document.getElementById('carte-blanche-root')
+  );
+};
